% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{character_ids}
\alias{character_ids}
\title{Find unique ID for all characters in the Marvel Universe.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1561 rows and 2 columns.
}
\usage{
character_ids
}
\description{
Find unique ID for all characters in the Marvel Universe.
}
\keyword{datasets}
