% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comics_id.R, R/data.R
\docType{data}
\name{comics_id}
\alias{comics_id}
\title{Get unique ID for Marvel Comics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49361 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49361 rows and 2 columns.
}
\usage{
comics_id

comics_id
}
\description{
Get unique ID for Marvel Comics

Get unique ID for all Marvel Comics.
}
\keyword{datasets}
