% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creators_id}
\alias{creators_id}
\title{Find unique ID of all creators in the Marvel Universe.}
\format{
A \code{tibble} containing 1,561 observations of two columns:
\describe{
\item{id}{A numeric vector containing the unique identifier associated with
each creator}
\item{fullName}{A character vector containing the name of each Marvel creator}
}
}
\source{
Data provided by Marvel. Copyright 2014 Marvel.
}
\usage{
creators_id
}
\description{
Contains the unique id and name for every Marvel creator whose
information can be accessed using the Marvel API.
}
\keyword{datasets}
