% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events_id}
\alias{events_id}
\title{Find unique ID for all important events in the Marvel Universe.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 3 columns.
}
\usage{
events_id
}
\description{
Find unique ID for all important events in the Marvel Universe.
}
\keyword{datasets}
