% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characters.R
\name{get_characters}
\alias{get_characters}
\title{Fetches a dataframe of Marvel characters from the Marvel API.}
\usage{
get_characters(
  limit = 100,
  offset = 0,
  comic = NULL,
  event = NULL,
  series = NULL
)
}
\arguments{
\item{limit}{An integer. Limit the result set to the specified number of resources (max 100).}

\item{offset}{An integer. Skip the specified number of resources in the result set.}

\item{comic}{A character vector. A Marvel comic issue, collection, graphic novel, or digital comic}

\item{event}{A character vector. A big, universe-changing storyline}

\item{series}{A character vector. Sequentially numbered list of comics with the same title and volume}

\item{nameStartsWith}{A character. Return characters with names that begin with the specified string (e.g. Sp).}
}
\value{
A dataframe of up to 100 characters containing their ID, name, description, time last modified, resource URI, and nested lists of thumbnails, comics, series, stories, events, and urls.
}
\description{
Fetches a dataframe of Marvel characters from the Marvel API using different user-set parameters. By default, the function will return the first 100 Marvel characters alphabetically. However, user can use different arguments in the function to find characters from different comics, events, stories, or series, or other arguments such as what the character's name starts with or how many results to offset by.
}
\examples{

}
\author{
Bethany Leap, Daniel Bernal Panqueva, Dashiell Nusbaum, Ian M Davis
}
