% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comics.R
\name{get_comics}
\alias{get_comics}
\title{Get tibble of comics information from Marvel API}
\usage{
get_comics(
  limit = 100,
  offset = 0,
  creator = NULL,
  event = NULL,
  series = NULL
)
}
\arguments{
\item{limit}{An integer. Limit the result set to the specified number of resources (defaulted to 100).}

\item{offset}{An integer. Skip the specified number of resources in the result set.}

\item{creator}{A character vector. A person or entity that make comics.}

\item{event}{A character vector. A big, universe-changing storyline}

\item{series}{A character vector. Sequentially numbered list of comics with the same title and volume}
}
\description{
Get tibble containing information related to comics through the Marvel API based on user input criteria.
}
\examples{

get_comics(creator = "Mary H.K. Choi")


get_comics(event = "Civil War")
}
\author{
Bethany Leap, Daniel Bernal Panqueva, Dashiell Nusbaum, Ian M Davis
}
