% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_creators.R
\name{get_creators}
\alias{get_creators}
\title{Get the Marvel creators}
\usage{
get_creators(public_key = NULL, private_key = NULL, name = NULL)
}
\arguments{
\item{key}{A string containing your}
}
\description{
This function accesses the Marvel API, gets the list of creators, and returns them in a tidy dataframe. Alternatively, give a string name and return all information about that creator in a tidy format.
}
