% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get tibble of events information from Marvel API}
\usage{
get_events(
  limit = 100,
  offset = 0,
  comic = NULL,
  creator = NULL,
  series = NULL,
  character = NULL
)
}
\arguments{
\item{limit}{An integer. Limit the result set to the specified number of resources (defaulted to 100).}

\item{offset}{An integer. Skip the specified number of resources in the result set.}

\item{comic}{A character vector. A Marvel comic issue, collection, graphic novel, or digital comic.}

\item{creator}{A character vector. A person or entity that makes comics.}

\item{series}{A character vector. Sequentially numbered list of comics with the same title and volume.}

\item{character}{A character vector. Full name of the Marvel character.}
}
\description{
Get tibble containing information related to events through the Marvel API based on user input criteria.
}
\examples{
get_events()
                
get_events(creator = "Brian Michael Bendis")

get_events(character = "Black Widow")


}
\author{
Bethany Leap, Daniel Bernal Panqueva, Dashiell Nusbaum, Ian M Davis
}
