% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{Get tibble of series information from Marvel API}
\usage{
get_series(
  limit = 100,
  offset = 0,
  comic = NULL,
  creator = NULL,
  events = NULL,
  character = NULL,
  story = NULL
)
}
\arguments{
\item{limit}{An integer. Limit the result set to the specified number of resources (defaulted to 100).}

\item{offset}{An integer. Skip the specified number of resources in the result set.}

\item{comic}{A character vector. A Marvel comic issue, collection, graphic novel, or digital comic.}

\item{creator}{A character vector. A person or entity that makes comics.}

\item{events}{A character vector. A big, universe-changing story line.}

\item{character}{A character vector. Full name of the Marvel character.}
}
\description{
Get tibble containing information related to series through the Marvel API based on user input criteria.
}
\examples{
get_series()

get_series(creator = "Aaron Alexovich")

get_series(events = "Age of Ultron")

get_series(character = "Vision")



}
\author{
Bethany Leap, Daniel Bernal Panqueva, Dashiell Nusbaum, Ian M Davis
}
