% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stories.R
\name{get_stories}
\alias{get_stories}
\title{Get tibble of stories information from Marvel API}
\usage{
get_stories(
  limit = 100,
  offset = 0,
  comic = NULL,
  creator = NULL,
  events = NULL,
  character = NULL,
  series = NULL
)
}
\arguments{
\item{limit}{An integer. Limit the result set to the specified number of resources (defaulted to 100).}

\item{offset}{An integer. Skip the specified number of resources in the result set.}

\item{comic}{A character vector. A Marvel comic issue, collection, graphic novel, or digital comic.}

\item{creator}{A character vector. A person or entity that makes comics.}

\item{events}{A character vector. A big, universe-changing story line.}

\item{character}{A character vector. Full name of the Marvel character.}

\item{series}{A character vector. Sequentially numbered list of comics with the same title and volume.}
}
\description{
Get tibble containing information related to stories through the Marvel API based on user input criteria.
}
\examples{
get_stories()

get_stories(comic = "A-Force (2015) #2")

get_stories(creator = "Andres Coelho")

get_stories(character = "Invisible Woman")

}
\author{
Bethany Leap, Daniel Bernal Panqueva, Dashiell Nusbaum, Ian M Davis
}
