% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{marvel_api}
\alias{marvel_api}
\title{Access the base point for the Marvel API}
\usage{
marvel_api(path, public_key, private_key)
}
\arguments{
\item{path}{the end point}
}
\description{
This is the base function for accessing the Marvel API base point. It is the foundation of all functions in the package.
}
\author{
Bethany Leap
}
