% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{marvel_api_key}
\alias{marvel_api_key}
\title{Save a Marvel API key to the \code{.Renviron} for future use.}
\usage{
marvel_api_key(key, overwrite = FALSE, create = FALSE)
}
\arguments{
\item{key}{A string containing the}
}
\description{
This function allows the user to save their Marvel API key securely in their
\code{.Renviron} file. Only run this in the console, do not save in your script.
}
\author{
Bethany Leap
}
