% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marvel_api_keys.R
\name{marvel_api_keys}
\alias{marvel_api_keys}
\title{Allows user to set public and private API keys}
\usage{
marvel_api_keys(
  public_api_key,
  private_api_key,
  overwrite = FALSE,
  install = FALSE
)
}
\arguments{
\item{public_api_key}{The public API key for the Marvel API, formatted in quotes (""). You can obtain a key from https://developer.marvel.com/account.}

\item{private_api_key}{The private API key for the Marvel API, formatted in quotes (""). You can obtain a key from https://developer.marvel.com/account.}

\item{overwrite}{If TRUE, overwrite the existing MARVEL_PUBLIC_API_KEY and MARVEL_PRIVATE_API_KEY in your .Renviron file, if FALSE (default), do not.}

\item{install}{If TRUE, install the key in your .Renviron file to be used in future sessions, if FALSE (default), do not.}
}
\description{
Allows user to easily set their public and private API keys for the Marvel API. Function allows the user to either set these keys for the current R session (in which case this function will need to be used every session in which the user wants to use this package) or add these keys to their R environment (the .Renviron file) in which case they would not need to use this function more than once or in future R sessions. Adding these keys to the .Renviron file allows the user to use other functions in this package without having their API keys exposed anywhere in their code. If the user does not have a .Renviron file, then this function creates a .Renviron file for them. If the user does have the file, this function will add these keys to the file and make a backup of the original file.
}
\section{References}{

Used the {tidycensus} package, particularly the census_api_key() function, as a framework for creating this function.
}

\examples{

 \dontrun{
marvel_api_keys(public_api_key = "abcd1234",
                private_api_key = "efgh5678",
                overwrite = TRUE,
                install = TRUE)
                }

}
\author{
Bethany Leap, Daniel Bernal Panqueva, Dashiell Nusbaum, Ian M Davis
}
