% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{character_ids}
\alias{character_ids}
\title{Find unique ID for all characters in the Marvel Universe.}
\format{
A \code{tibble} containing 1,561 observations of two columns:
\describe{
\item{id}{A numeric vector containing the unique identifier associated with
each character}
\item{name}{A character vector containing the name of each Marvel character}
}
}
\source{
Data provided by Marvel. Copyright 2014 Marvel.
}
\usage{
character_ids
}
\description{
Contains the unique id and name for every Marvel character which
can be called using the Marvel API.
}
\keyword{datasets}
