% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comics_id}
\alias{comics_id}
\title{Get unique ID for all Marvel Comics.}
\format{
a \code{tibble} containing 49,361 observations of 2 columns:
\describe{
\item{id}{A numeric vector containing the unique identifier associated with
each comic}
\item{title}{A character vector containing the title of each Marvel comic}
}
}
\source{
Data provided by Marvel. Copyright 2014 Marvel.
}
\usage{
comics_id
}
\description{
Contains the unique ID and title of all comics which can be
accessed using the Marvel API.
}
\keyword{datasets}
