% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creators_id}
\alias{creators_id}
\title{Find unique ID of all creators in the Marvel Universe.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5693 rows and 2 columns.
}
\usage{
creators_id
}
\description{
Find unique ID of all creators in the Marvel Universe.
}
\keyword{datasets}
