% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events_id}
\alias{events_id}
\title{Find unique ID for all important events in the Marvel Universe.}
\format{
A \code{data.frame} with 74 observations and 3 columns:
\describe{
\item{id}{The unique identifier associated with each event}
\item{title}{The title of the important event}
\item{description}{}
}
}
\usage{
events_id
}
\description{
Find unique ID for all important events in the Marvel Universe.
}
\keyword{datasets}
