% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_creators.R
\name{get_creators}
\alias{get_creators}
\title{Get tibble of creators information from Marvel API}
\usage{
get_creators(
  limit = 100,
  offset = 0,
  comic = NULL,
  event = NULL,
  series = NULL
)
}
\arguments{
\item{limit}{An integer. Limit the result set to the specified number of resources (defaulted to 100).}

\item{offset}{An integer. Skip the specified number of resources in the result set.}

\item{comic}{A character vector. A Marvel comic issue, collection, graphic novel, or digital comic}

\item{event}{A character vector. A big, universe-changing storyline}

\item{series}{A character vector. Sequentially numbered list of comics with the same title and volume}
}
\description{
Get tibble containing information related to creators through the Marvel API based on user input criteria.
}
\examples{
get_creators(limit = 50,
                offset = 0,
                comic = "X-Men (1991) #23")
                
                
get_creators(event = "Civil War") 
}
\author{
Bethany Leap, Daniel Bernal Panqueva, Dashiell Nusbaum, Ian M Davis
}
