% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{series_id}
\alias{series_id}
\title{Find unique ID for all series in the Marvel Universe.}
\format{
a \code{tibble} containing 9,309 observations of two columns:
\describe{
\item{id}{A numeric vector containing the unique identifier associated with
each Marvel series}
\item{title}{A character vector containing the title of each Marvel series}
}
}
\source{
Data provided by Marvel. Copyright 2014 Marvel.
}
\usage{
series_id
}
\description{
Contains the unique id and title for every Marvel series which
can be called using the Marvel API.
}
\keyword{datasets}
