% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_id.R
\docType{data}
\name{series_id}
\alias{series_id}
\title{Find unique ID for series in the Marvel Universe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9309 rows and 2 columns.
}
\usage{
series_id
}
\description{
Find unique ID for series in the Marvel Universe
}
\keyword{datasets}
